Move blockchain.* functions
Move blockchain.* functions from mod.rs -> blockchain.rs
