+++ src/rpc/mod.rs
@@ -1,10 +1,6 @@
 use bitcoin::blockdata::transaction::Transaction;
-use bitcoin::consensus::encode::{deserialize, serialize};
-use bitcoin_hashes::hex::ToHex;
 use bitcoin_hashes::sha256d::Hash as Sha256dHash;
-use bitcoin_hashes::Hash;
 use error_chain::ChainedError;
-use hex;
 use serde_json::{from_str, Value};
 use std::collections::{HashMap, HashSet};
 use std::io::{BufRead, BufReader, Write};
@@ -16,12 +12,10 @@ use std::time::Duration;
 
 use crate::def::PROTOCOL_VERSION_MAX;
 use crate::errors::*;
-use crate::mempool::MEMPOOL_HEIGHT;
-use crate::metrics::{Gauge, HistogramOpts, HistogramVec, MetricOpts, Metrics};
-use crate::query::{Query, Status};
-use crate::rpc::parseutil::{
-    bool_from_value_or, hash_from_value, rpc_arg_error, usize_from_value, usize_from_value_or,
-};
+use crate::metrics::{HistogramOpts, HistogramVec, Metrics};
+use crate::query::Query;
+use crate::rpc::blockchain::BlockchainRPC;
+use crate::rpc::parseutil::usize_from_value;
 use crate::rpc::server::{
     server_add_peer, server_banner, server_donation_address, server_features,
     server_peers_subscribe, server_version,
@@ -31,23 +25,9 @@ use crate::timeout::TimeoutTrigger;
 use crate::util::FullHash;
 use crate::util::{spawn_thread, Channel, HeaderEntry, SyncChannel};
 
+pub mod blockchain;
 pub mod parseutil;
 pub mod server;
-
-fn unspent_from_status(status: &Status) -> Value {
-    json!(Value::Array(
-        status
-            .unspent()
-            .into_iter()
-            .map(|out| json!({
-                "height": out.height,
-                "tx_pos": out.output_index,
-                "tx_hash": out.txn_id.to_hex(),
-                "value": out.value,
-            }))
-            .collect()
-    ))
-}
 
 fn get_output_scripthash(txn: &Transaction, n: Option<usize>) -> Vec<FullHash> {
     if let Some(out) = n {
@@ -62,19 +42,18 @@ fn get_output_scripthash(txn: &Transaction, n: Option<usize>) -> Vec<FullHash> {
 
 struct Connection {
     query: Arc<Query>,
-    last_header_entry: Option<HeaderEntry>,
-    status_hashes: HashMap<Sha256dHash, Value>, // ScriptHash -> StatusHash
     stream: TcpStream,
     addr: SocketAddr,
     chan: SyncChannel<Message>,
     stats: Arc<Stats>,
-    relayfee: f64,
     rpc_timeout: u16,
+    blockchainrpc: BlockchainRPC,
 }
 
 impl Connection {
     pub fn new(
         query: Arc<Query>,
+        metrics: Arc<Metrics>,
         stream: TcpStream,
         addr: SocketAddr,
         stats: Arc<Stats>,
@@ -83,290 +62,18 @@ impl Connection {
         buffer_size: usize,
     ) -> Connection {
         Connection {
-            query,
-            last_header_entry: None, // disable header subscription for now
-            status_hashes: HashMap::new(),
+            query: query.clone(),
             stream,
             addr,
             chan: SyncChannel::new(buffer_size),
-            stats,
-            relayfee,
+            stats: stats.clone(),
             rpc_timeout,
+            blockchainrpc: BlockchainRPC::new(query, metrics, relayfee, rpc_timeout),
         }
-    }
-
-    fn blockchain_headers_subscribe(&mut self) -> Result<Value> {
-        let entry = self.query.get_best_header()?;
-        let hex_header = hex::encode(serialize(entry.header()));
-        let result = json!({"hex": hex_header, "height": entry.height()});
-        self.last_header_entry = Some(entry);
-        Ok(result)
     }
 
     fn mempool_get_fee_histogram(&self) -> Result<Value> {
         Ok(json!(self.query.get_fee_histogram()))
-    }
-
-    fn blockchain_block_header(&self, params: &[Value]) -> Result<Value> {
-        let height = usize_from_value(params.get(0), "height")?;
-        let cp_height = usize_from_value_or(params.get(1), "cp_height", 0)?;
-
-        let raw_header_hex: String = self
-            .query
-            .get_headers(&[height])
-            .into_iter()
-            .map(|entry| hex::encode(&serialize(entry.header())))
-            .collect();
-
-        if cp_height == 0 {
-            return Ok(json!(raw_header_hex));
-        }
-        let (branch, root) = self.query.get_header_merkle_proof(height, cp_height)?;
-
-        let branch_vec: Vec<String> = branch.into_iter().map(|b| b.to_hex()).collect();
-
-        Ok(json!({
-            "header": raw_header_hex,
-            "root": root.to_hex(),
-            "branch": branch_vec
-        }))
-    }
-
-    fn blockchain_block_headers(&self, params: &[Value]) -> Result<Value> {
-        let start_height = usize_from_value(params.get(0), "start_height")?;
-        let count = usize_from_value(params.get(1), "count")?;
-        let cp_height = usize_from_value_or(params.get(2), "cp_height", 0)?;
-        let heights: Vec<usize> = (start_height..(start_height + count)).collect();
-        let headers: Vec<String> = self
-            .query
-            .get_headers(&heights)
-            .into_iter()
-            .map(|entry| hex::encode(&serialize(entry.header())))
-            .collect();
-
-        if count == 0 || cp_height == 0 {
-            return Ok(json!({
-                "count": headers.len(),
-                "hex": headers.join(""),
-                "max": 2016,
-            }));
-        }
-
-        let (branch, root) = self
-            .query
-            .get_header_merkle_proof(start_height + (count - 1), cp_height)?;
-
-        let branch_vec: Vec<String> = branch.into_iter().map(|b| b.to_hex()).collect();
-
-        Ok(json!({
-            "count": headers.len(),
-            "hex": headers.join(""),
-            "max": 2016,
-            "root": root.to_hex(),
-            "branch" : branch_vec
-        }))
-    }
-
-    fn blockchain_estimatefee(&self, params: &[Value]) -> Result<Value> {
-        let blocks_count = usize_from_value(params.get(0), "blocks_count")?;
-        let fee_rate = self.query.estimate_fee(blocks_count); // in BTC/kB
-        Ok(json!(fee_rate.max(self.relayfee)))
-    }
-
-    fn blockchain_relayfee(&self) -> Result<Value> {
-        Ok(json!(self.relayfee)) // in BTC/kB
-    }
-
-    fn blockchain_scripthash_subscribe(
-        &mut self,
-        params: &[Value],
-        timeout: &TimeoutTrigger,
-    ) -> Result<Value> {
-        let script_hash = hash_from_value(params.get(0)).chain_err(|| "bad script_hash")?;
-        let status = self.query.status(&script_hash[..], timeout)?;
-        let result = status.hash().map_or(Value::Null, |h| json!(hex::encode(h)));
-        self.status_hashes.insert(script_hash, result.clone());
-        self.stats
-            .subscriptions
-            .set(self.status_hashes.len() as i64);
-        Ok(result)
-    }
-
-    fn blockchain_scripthash_get_balance(
-        &self,
-        params: &[Value],
-        timeout: &TimeoutTrigger,
-    ) -> Result<Value> {
-        let script_hash = hash_from_value(params.get(0))?;
-        let status = self.query.status(&script_hash[..], timeout)?;
-        Ok(
-            json!({ "confirmed": status.confirmed_balance(), "unconfirmed": status.mempool_balance() }),
-        )
-    }
-
-    fn blockchain_scripthash_get_history(
-        &self,
-        params: &[Value],
-        timeout: &TimeoutTrigger,
-    ) -> Result<Value> {
-        let script_hash = hash_from_value(params.get(0))?;
-        let status = self.query.status(&script_hash[..], timeout)?;
-        Ok(json!(Value::Array(
-            status
-                .history()
-                .into_iter()
-                .map(|item| json!({"height": item.0, "tx_hash": item.1.to_hex()}))
-                .collect()
-        )))
-    }
-
-    fn blockchain_scripthash_listunspent(
-        &self,
-        params: &[Value],
-        timeout: &TimeoutTrigger,
-    ) -> Result<Value> {
-        let script_hash = hash_from_value(params.get(0))?;
-        Ok(unspent_from_status(
-            &self.query.status(&script_hash[..], timeout)?,
-        ))
-    }
-
-    fn blockchain_scripthash_get_first_use(&self, params: &[Value]) -> Result<Value> {
-        let scripthash = hash_from_value(params.get(0)).chain_err(|| "bad script_hash")?;
-
-        let to_fullhash = |h: &Sha256dHash| -> [u8; 32] {
-            let mut res = [0; 32];
-            res[..].copy_from_slice(&h[..]);
-            res
-        };
-
-        let firstuse = self.query.scripthash_first_use(&to_fullhash(&scripthash))?;
-        if firstuse.0 == 0 {
-            return Err(ErrorKind::RpcError(
-                RpcErrorCode::NotFound,
-                format!("scripthash '{}' not found", scripthash.to_hex()),
-            )
-            .into());
-        }
-        let hash = if firstuse.0 == MEMPOOL_HEIGHT {
-            Sha256dHash::default()
-        } else {
-            let h = self.query.get_headers(&[firstuse.0 as usize]);
-            if h.is_empty() {
-                warn!("expected to find header for height {}", firstuse.0);
-                Sha256dHash::default()
-            } else {
-                *h[0].hash()
-            }
-        };
-
-        Ok(json!({
-            "block_hash": hash.to_hex(),
-            "block_height": if firstuse.0 == MEMPOOL_HEIGHT { 0 } else { firstuse.0 },
-            "tx_hash": firstuse.1.to_hex()
-        }))
-    }
-
-    fn blockchain_transaction_broadcast(&self, params: &[Value]) -> Result<Value> {
-        let tx = params.get(0).chain_err(|| rpc_arg_error("missing tx"))?;
-        let tx = tx.as_str().chain_err(|| rpc_arg_error("non-string tx"))?;
-        let tx = hex::decode(&tx).chain_err(|| rpc_arg_error("non-hex tx"))?;
-        let tx: Transaction = deserialize(&tx).chain_err(|| rpc_arg_error("failed to parse tx"))?;
-        let txid = self
-            .query
-            .broadcast(&tx)
-            .chain_err(|| rpc_arg_error("rejected by network"))?;
-        Ok(json!(txid.to_hex()))
-    }
-
-    fn blockchain_transaction_get(&self, params: &[Value]) -> Result<Value> {
-        let tx_hash = hash_from_value(params.get(0))?;
-        let verbose = match params.get(1) {
-            Some(value) => value.as_bool().chain_err(|| "non-bool verbose value")?,
-            None => false,
-        };
-        if !verbose {
-            let tx = self.query.load_txn(&tx_hash, None, None)?;
-            Ok(json!(hex::encode(serialize(&tx))))
-        } else {
-            let header = self.query.lookup_blockheader(&tx_hash, None)?;
-            let blocktime = match header {
-                Some(ref header) => header.header().time,
-                None => 0,
-            };
-            let height = match header {
-                Some(ref header) => header.height(),
-                None => 0,
-            };
-            let confirmations = match header {
-                Some(ref header) => {
-                    if let Some(best) = self.query.best_header() {
-                        best.height() - header.height()
-                    } else {
-                        0
-                    }
-                }
-                None => 0,
-            };
-            let blockhash = header.and_then(|h| Some(h.hash().clone()));
-            let tx = self.query.load_txn(&tx_hash, blockhash.as_ref(), None)?;
-
-            let tx_serialized = serialize(&tx);
-            Ok(json!({
-                "blockhash": blockhash.unwrap_or(Sha256dHash::default()).to_hex(),
-                "blocktime": blocktime,
-                "height": height,
-                "confirmations": confirmations,
-                "hash": tx.txid().to_hex(),
-                "txid": tx.txid().to_hex(),
-                "size": tx_serialized.len(),
-                "hex": hex::encode(tx_serialized),
-                "locktime": tx.lock_time,
-                "time": blocktime,
-                "vin": tx.input.iter().map(|txin| json!({
-                    "sequence": txin.sequence,
-                    "txid": txin.previous_output.txid.to_hex(),
-                    "vout": txin.previous_output.vout,
-                    "scriptSig": txin.script_sig.to_hex(),
-                })).collect::<Vec<Value>>(),
-                "vout": tx.output.iter().map(|txout| json!({
-                    "value": txout.value,
-                    "scriptPubKey": txout.script_pubkey.to_hex()
-                })).collect::<Vec<Value>>(),
-            }))
-        }
-    }
-
-    fn blockchain_transaction_get_merkle(&self, params: &[Value]) -> Result<Value> {
-        let tx_hash = hash_from_value(params.get(0))?;
-        let height = usize_from_value(params.get(1), "height")?;
-        let (merkle, pos) = self
-            .query
-            .get_merkle_proof(&tx_hash, height)
-            .chain_err(|| "cannot create merkle proof")?;
-        let merkle: Vec<String> = merkle.into_iter().map(|txid| txid.to_hex()).collect();
-        Ok(json!({
-                "block_height": height,
-                "merkle": merkle,
-                "pos": pos}))
-    }
-
-    fn blockchain_transaction_id_from_pos(&self, params: &[Value]) -> Result<Value> {
-        let height = usize_from_value(params.get(0), "height")?;
-        let tx_pos = usize_from_value(params.get(1), "tx_pos")?;
-        let want_merkle = bool_from_value_or(params.get(2), "merkle", false)?;
-
-        let (txid, merkle) = self.query.get_id_from_pos(height, tx_pos, want_merkle)?;
-
-        if !want_merkle {
-            return Ok(json!(txid.to_hex()));
-        }
-
-        let merkle_vec: Vec<String> = merkle.into_iter().map(|entry| entry.to_hex()).collect();
-
-        Ok(json!({
-            "tx_hash" : txid.to_hex(),
-            "merkle" : merkle_vec}))
     }
 
     fn cashaccount_query_name(&self, params: &[Value]) -> Result<Value> {
@@ -385,31 +92,33 @@ impl Connection {
             .start_timer();
         let timeout = TimeoutTrigger::new(Duration::from_secs(self.rpc_timeout as u64));
         let result = match method {
-            "blockchain.block.header" => self.blockchain_block_header(&params),
-            "blockchain.block.headers" => self.blockchain_block_headers(&params),
-            "blockchain.estimatefee" => self.blockchain_estimatefee(&params),
-            "blockchain.headers.subscribe" => self.blockchain_headers_subscribe(),
-            "blockchain.relayfee" => self.blockchain_relayfee(),
+            "blockchain.block.header" => self.blockchainrpc.block_header(&params),
+            "blockchain.block.headers" => self.blockchainrpc.block_headers(&params),
+            "blockchain.estimatefee" => self.blockchainrpc.estimatefee(&params),
+            "blockchain.headers.subscribe" => self.blockchainrpc.headers_subscribe(),
+            "blockchain.relayfee" => self.blockchainrpc.relayfee(),
             "blockchain.scripthash.get_balance" => {
-                self.blockchain_scripthash_get_balance(&params, &timeout)
-            }
-            "blockchain.scripthash.get_history" => {
-                self.blockchain_scripthash_get_history(&params, &timeout)
-            }
-            "blockchain.scripthash.listunspent" => {
-                self.blockchain_scripthash_listunspent(&params, &timeout)
-            }
-            "blockchain.scripthash.subscribe" => {
-                self.blockchain_scripthash_subscribe(&params, &timeout)
+                self.blockchainrpc.scripthash_get_balance(&params, &timeout)
             }
             "blockchain.scripthash.get_first_use" => {
-                self.blockchain_scripthash_get_first_use(&params)
+                self.blockchainrpc.scripthash_get_first_use(&params)
             }
-            "blockchain.transaction.broadcast" => self.blockchain_transaction_broadcast(&params),
-            "blockchain.transaction.get" => self.blockchain_transaction_get(&params),
-            "blockchain.transaction.get_merkle" => self.blockchain_transaction_get_merkle(&params),
+            "blockchain.scripthash.get_history" => {
+                self.blockchainrpc.scripthash_get_history(&params, &timeout)
+            }
+            "blockchain.scripthash.listunspent" => {
+                self.blockchainrpc.scripthash_listunspent(&params, &timeout)
+            }
+            "blockchain.scripthash.subscribe" => {
+                self.blockchainrpc.scripthash_subscribe(&params, &timeout)
+            }
+            "blockchain.transaction.broadcast" => self.blockchainrpc.transaction_broadcast(&params),
+            "blockchain.transaction.get" => self.blockchainrpc.transaction_get(&params),
+            "blockchain.transaction.get_merkle" => {
+                self.blockchainrpc.transaction_get_merkle(&params)
+            }
             "blockchain.transaction.id_from_pos" => {
-                self.blockchain_transaction_id_from_pos(&params)
+                self.blockchainrpc.transaction_id_from_pos(&params)
             }
             "mempool.get_fee_histogram" => self.mempool_get_fee_histogram(),
             "server.add_peer" => server_add_peer(),
@@ -464,67 +173,7 @@ impl Connection {
         })
     }
 
-    fn on_scripthash_change(&mut self, scripthash: FullHash) -> Result<()> {
-        let scripthash = Sha256dHash::from_slice(&scripthash[..]).expect("invalid scripthash");
-
-        let old_statushash;
-        match self.status_hashes.get(&scripthash) {
-            Some(statushash) => {
-                old_statushash = statushash;
-            }
-            None => {
-                return Ok(());
-            }
-        };
-
-        let timer = self
-            .stats
-            .latency
-            .with_label_values(&["statushash_update"])
-            .start_timer();
-
-        let timeout = TimeoutTrigger::new(Duration::from_secs(self.rpc_timeout as u64));
-        let status = self.query.status(&scripthash[..], &timeout)?;
-        let new_statushash = status.hash().map_or(Value::Null, |h| json!(hex::encode(h)));
-        if new_statushash == *old_statushash {
-            return Ok(());
-        }
-        timer.observe_duration();
-        self.send_values(&vec![json!({
-                    "jsonrpc": "2.0",
-                    "method": "blockchain.scripthash.subscribe",
-                    "params": [scripthash.to_hex(), new_statushash]})])?;
-        self.status_hashes.insert(scripthash, new_statushash);
-        Ok(())
-    }
-
-    fn on_chaintip_change(&mut self, chaintip: HeaderEntry) -> Result<()> {
-        let timer = self
-            .stats
-            .latency
-            .with_label_values(&["chaintip_update"])
-            .start_timer();
-
-        if let Some(ref mut last_entry) = self.last_header_entry {
-            if *last_entry == chaintip {
-                return Ok(());
-            }
-
-            *last_entry = chaintip;
-            let hex_header = hex::encode(serialize(last_entry.header()));
-            let header = json!({"hex": hex_header, "height": last_entry.height()});
-            self.send_values(&vec![
-                (json!({
-                "jsonrpc": "2.0",
-                "method": "blockchain.headers.subscribe",
-                "params": [header]})),
-            ])?;
-        }
-        timer.observe_duration();
-        Ok(())
-    }
-
-    fn send_values(&mut self, values: &[Value]) -> Result<()> {
+    pub fn send_values(&mut self, values: &[Value]) -> Result<()> {
         for value in values {
             let line = value.to_string() + "\n";
             if let Err(e) = self.stream.write_all(line.as_bytes()) {
@@ -557,8 +206,18 @@ impl Connection {
                     };
                     self.send_values(&[reply])?
                 }
-                Message::ScriptHashChange(hash) => self.on_scripthash_change(hash)?,
-                Message::ChainTipChange(tip) => self.on_chaintip_change(tip)?,
+                Message::ScriptHashChange(hash) => {
+                    let notification = self.blockchainrpc.on_scripthash_change(hash)?;
+                    if let Some(n) = notification {
+                        self.send_values(&[n])?;
+                    }
+                }
+                Message::ChainTipChange(tip) => {
+                    let notification = self.blockchainrpc.on_chaintip_change(tip)?;
+                    if let Some(n) = notification {
+                        self.send_values(&[n])?;
+                    }
+                }
                 Message::Done => return Ok(()),
             }
         }
@@ -633,7 +292,6 @@ pub struct RPC {
 
 struct Stats {
     latency: HistogramVec,
-    subscriptions: Gauge,
 }
 
 impl RPC {
@@ -690,7 +348,7 @@ impl RPC {
     pub fn start(
         addr: SocketAddr,
         query: Arc<Query>,
-        metrics: &Metrics,
+        metrics: Arc<Metrics>,
         relayfee: f64,
         rpc_timeout: u16,
         rpc_buffer_size: usize,
@@ -700,10 +358,6 @@ impl RPC {
                 HistogramOpts::new("electrscash_electrum_rpc", "Electrum RPC latency (seconds)"),
                 &["method"],
             ),
-            subscriptions: metrics.gauge(MetricOpts::new(
-                "electrscash_electrum_subscriptions",
-                "# of Electrum subscriptions",
-            )),
         });
         let notification = Channel::unbounded();
         RPC {
@@ -729,11 +383,13 @@ impl RPC {
                         let senders = Arc::clone(&senders);
                         let stats = Arc::clone(&stats);
                         let handles = Arc::clone(&handles);
+                        let metrics = Arc::clone(&metrics);
 
                         spawn_thread("peer", move || {
                             info!("[{}] connected peer #{}", addr, handle_id);
                             let conn = Connection::new(
                                 query,
+                                metrics,
                                 stream,
                                 addr,
                                 stats,
